requirements.txt
faker
sqlalchemy
pyyaml
psycopg2-binary
boto3
pandas
pyarrow
google-cloud-storage
apache-beam
python-dotenv